Add script to get roles from NW server
Adding a script to query the mongo database on the NetWitness server and return a json list of roles with their associated permissions. Will further refine this json list into our excel template later.

Also begin work on a script to get health stats from the NW services and post that info to Splunk. Work-in-progress at this time