Optimize Get-Roles
Removed the need for the separate bash script and javascript file on the NetWitness side as well as the additional role with all enabled permissions.
New version enumerates the list of all permissions from a different collection in Mongo and then uses that list directly to compare against all roles.